Proyecto_Inv_Server/requirements.txt
asgiref==3.11.0
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
Django==5.2.5
djangorestframework==3.16.1
exceptiongroup==1.3.1
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
mypy_extensions==1.1.0
mysql-connector-python==9.6.0
mysqlclient==2.2.7
numpy==1.26.4
outcome==1.3.0.post0
packaging==26.0
pandas==2.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
rarfile==4.2
selenium==4.40.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
tqdm==4.67.1
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
xmltodict==1.0.2
zipp==3.23.0
